Incorporated the create-fm goal into the include-artifact goal
Also added the ability to include additional features like repoinit etc
