+++ src/it/include-artifact-including-source-features/invoker.properties
