+++ src/it/include-artifact-including-source-features/src/main/features/feature-test-repoinit.json
@@ -1,0 +1,5 @@
+{
+  "repoinit:TEXT|true":[
+    "create path (rep:AuthorizableFolder) /home/users/system"
+  ]
+}
